requirements.txt
Pillow>=9.5.0
flask>=2.0.0
requests>=2.30.0
python-dotenv>=0.19.0 